fix(server): remove the request body limit on the PUT metadata
The maximum size of the metadata is actually always lower than the
default maximum request body limit (for reasonable number of
shards).

Further, it seems that the limit that was set (`metatadat_length +
headroom`) was not large enough, and caused storing metadata to fail
in some cases.

Reverting fixes this problem up to ~30k shards. If we plan to go
above that number, we may need to revisit.

Signed-off-by: giac-mysten <giacomo@mystenlabs.com>