feat(aggregator): reflect the content type header of the request
The aggregator reflects the `content-type` header of the requests,
when present.

Also, added the header `x-conte-type-options: nosniff` to the response
headers: This header stops the browser from loading `<script>` and
`<style>` elements if the MIME type is not the appropriate `text/css`
or a javascript MIME type. See:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Content-Type-Options#directives

This is intended as a way to prevent devs from unintentionally loading
octet streams from Walrus as executables.

We may want to revis this if it becomes clear that it is not needed,
or it is too restrictive.

Closes: #452

Signed-off-by: giac-mysten <giacomo@mystenlabs.com>