feat: add http tracing to the client
Add HTTP tracing to the client and propagate spans from the client to servers,
so that calls made using the client can be associated with traces at the server.

Additionally, rename functions in `walrus-sdk::client` such that generic methods
have no suffix and enum methods have a `_by_type` suffix.