feat(move) add register_thrifty function that minimizes storage_resource use
`blob::register` allows the storage resource to be "bigger" than the blob being registered:

```
assert!(encoded_size <= storage_size(&storage), EResourceSize);
```

This is harmless from the system's perspective, but potentially troubling for a user who uses the wrong storage resource to register a blob.

In addition to overpaying up front, `blob::extend` looks at the size of the storage resource locked in the blob, so a user who accidentally overpays once will also need to overpay upon extension.

This PR leaves the old `register` functionality in place for now, but adds a new `register_thrifty` function that splits the input storage resource into exactly the right size + returns the remainder to the user.