fix: clean up all `unwrap`s in production code
This PR removes all instances of `unwrap` in production code. This
excludes tests, benchmarks, and the `walrus-orchestrator` crate.

Some cases are replaced by an `expect` explaining the reason why this
can never panic. In some cases a slight rewrite of the code allowed
removing the unwrapping altogether.

There are two cases where the previous behavior was potentially
problematic:
- In `crates/walrus-event/src/checkpoint_processor.rs`
- The authorization header in
  `crates/walrus-service/src/node/server/extract.rs` and
  `crates/walrus-service/src/node/server/routes.rs`.