fix(client): avoid reading `Blob` object after certification
The only purpose of this was to check that the `certified_epoch` was
updated correctly. However, we *know* that this must have happened if
the `certify_blob` contract completed without error. We can therefore
simply modify the `Blob` object locally and safe one RPC request.

This speeds up the write process and avoids spurious errors in case of
load-balancing between Sui full nodes.