fix: adds a limit by the number of slivers being recovered
This allows limiting by the number of slivers being recovered. It
uses an iterator to create the sliver recovery futures, thereby avoiding
reserving space for all of the futures in advance. As permits are
available, it dispatches the future and awaits the result.

This does not resolve the issue of the amount of memory required for
syncing, however, it gives us another knob to tune for seeing the
performance of the current storage node deployment and maintaining
system stability.