refactor(node): make event handling more robust
This introduces a new `EventHandle` struct, which is used to track
handling of events and to mark them as complete. As this struct does not
implement `Clone`, it is impossible to mark the same event as complete
multiple times. Furthermore, it automatically logs an error if it is
dropped without being marked as complete.

No actual behavior is changed, but this should prevent programming
errors in the future.