[walrus/proxy]
## Description

create a walrus-proxy, similar to how we have a sui-proxy. Some items are borrowed, but there is a considerable amount of customization here as well.

General flow of how this works;

walrus-node spins up a reqwest client and connects to the walrus-proxy over anonymous TLS, accepting well known CA certs like we do today in other parts of our infra.
the network private key is used to sign a uuidv7 and that signature is sent in an Authorization header along with a message. The signature is a recoverable secp256r1 signature which allows the pub key to be recovered service side as long as the message that was signed is also known. We send the signature as base64 and also encode the uuidv7 as the message to be signed over as a string.
the walrus-proxy will inspect this header and extract the pub key from the signature, decoding as needed. If it matches a known walrus-node, we accept the metric. This authorization header can be swapped out later if we want mTLS, but there is a benefit in using the header mechanism because we can easily avoid tls termination on this service and instead rely on gcp's upstream load balancers to do this for us.
walrus-proxy side, we have a process that polls for a staking object, we extract committee members and we grab their pub key. this is key to how we know who to accept metrics from. This is done on a configurable interval and automatically adjusts to nodes joining or leaving the stacking object we depend on. If the staking object id were to change, then we'd need to update the variable that holds this object id. If this is a frequent occurrence, we could add an api endpoint to the service to reconfigure itself w/o a restart.
in a follow-up PR, I will add an additional layer of replay protection using a sliding window approach. The uuidv7 uses the most significant 48 bits to encode a unix epoch timestamp. We can order these uuids based on time and apply a replay constraint to middleware to prevent the same message from being used over and over again. Messages that fall outside of coordinated UTC windows will not be allowed, regardless of being seen prior.

## Test Plan

local tests

working as expected so far. I briefly had it on and captured some data. https://metrics.sui.io/goto/-R1hv-mHg?orgId=1

Closes #571