fix: correctly update initial certified epoch
We need to update the `initial_certified_epoch` every time there is a
period where the blob is not certified. This is important at the moment,
because we do not clean up expired blob data yet, but will also be
relevant afterwards, because the cleanup process will probably run in
the background, such that there are periods where the blob is no longer
certified but the data hasn't been cleaned up yet.