fix: increase number of permits for the thread pool
The prior amount was set to 1.5 the number of cores, which is only a few
to tens of dozens of requests depending on the machine. However, requests are
sent for batches of up to 100 symbols (in a 1000 shard setup). On a storage node
with 24 cores, this would mean that every request for 100 symbols, which would
attempt to concurrently acquire 100 permits, many symbols would simply fail to
get the permit.

In CI, this lead to the client failing to get sufficient recovery symbols after
a single round of attempts resulting in exponential backoff and test timeouts.

This PR fixes the issue by increasing the number of permits to a function of the
maximum amount of time that it's acceptable that a task is waiting in the queue.
For the same example of a host with 24 cores, the number of permits is now 2400
instead of of 36, which assumes that each permit is held for 10 ms and we want
requests to be complete within 1 second.