refactor(sdk): replace `StoreWhen` enum by `StoreOptimizations` struct
The current `StoreWhen` enum combines two orthogonal optimizations
(checking the current blob status and checking existing resources).
As a result, it needs four variants, and the exact implications of each
of them are a bit unclear.

This replaces this enum by a new `StoreOptimizations` struct, which
contains boolean fields for all (currently 2) optimizations, which can
be enabled independently.

No actual behavior is changed in the final binaries. There are some
minor changes in tests, which shouldn't affect the outcome, however.