feat(node): configurable sampled tracing interval with increased default
Previously, all the interval for all sampled tracing logs was hardcoded
to 30 seconds. This caused a lot of emitted logs that were not useful
during normal operation.

This adds a new configuration parameter for the event processor config
to control this interval. The default is increased to 1 hour.

Some additional changes:
- Allow providing `Duration` values directly in the `tracing_sampled`
  macro (in addition to literal humantime values).
- Enable lints in the `walrus-utils` crate.
- Fix some dependencies in the `walrus-utils` crate.
- Fix some log messages.
- Some other minor refactoring.