test: make unit tests work with `cargo test` and `cargo tarpaulin`
We normally run tests with `cargo nextest`, which has some additional
isolation mechanisms compared to `cargo test` and `cargo tarpaulin`.
This resulted in some tests not running successfully with `cargo test`
and `cargo tarpaulin`.

This adds some locks for some unit tests to prevent them from running
simultaneously in the same process. It also updates the tarpaulin config
and unifies how tracing subscribers are initialized in tests.

Unfortunately, some end-to-end tests still fail after this change. A
follow-up issue was created for this.