fix(walrus-sui): limit size of futures on the stack
https://github.com/MystenLabs/walrus/pull/2459 significantly increased
the size of some futures in the `walrus-sui` crate. This can result in
stack overflows, particularly for dev builds.

This enables the `clippy::large_futures` lint for the `walrus-sui` crate
to catch these cases and reduces the amount of stack space required for
some futures by boxing them in strategic places (as suggested by the
Clippy lint).

Also contains some clean up that was missed in
https://github.com/MystenLabs/walrus/pull/2459.