chore: remove tarpaulin (#2539)
Our coverage-report setup with tarpaulin has been broken for a long
time. It turned out to be pretty hard to get it working again and
maintaining it is also a lot of work.

This removes all related (unused) actions and docs. If we need test
coverage in the future, we can revisit this and set up a more robust
solution.