fix(core): optimize memory allocations during blob decoding
In the past, when we were using RaptorQ for the 1D encoding, it was
necessary to keep around intermediate decoding state as the exact number
of slivers needed for decoding was unknown and we would potentially have
to reuse the decoder for a new decoding attempt.

This is no longer the case with Reed-Solomon, as we know the exact number
of slivers needed for decoding. As a result, we can allocate a single 1D
decoder and reuse it for all sub-decodings during blob decoding.

This commit also makes some other changes to reduce the number of
allocations during blob decoding, specifically by using flat arrays
instead of nested vectors where possible.

Both together result in both a major reduction in memory usage and a
significant performance improvement:

- The peak memory usage while reading a 1GiB sliver from the public
  Testnet (before the blob re-encoding) was reduced from more than 10GiB
  to ~2.5GiB.
- The decoding time (without re-encoding) for 1GiB blobs was reduced by
  more than 60% (corresponding to a ~160% throughput improvement).