feat(client): add options to disable consistency check
This PR adds a CLI option `--skip-consistency-check` and an aggregator
query parameter `skip_consistency_check` to disable the consistency
check when reading a blob.

This is useful to reduce the latency when reading a blob that was stored
by oneself or someone else who is known and trusted. When reading a
10MiB blob through a local aggregator from Testnet, the latency is
reduced from ~1.2s to ~0.7s.

Also contains the following changes:
- Refactor some code in `walrus-core`.
- Prepare the code for adding the new default consistency check.
- Fix an inappropriate panic in the client when reading inconsistent
  blobs.
- Improve the documentation of the consistency check.
- Improve the blob encoding benchmarks to test the different consistency
  checks.