feat(client): implement new default consistency check
This adds a new default consistency check that only verifies the hashes
of the systematic primary slivers instead of doing a full re-encoding.
This is significantly faster and only slightly reduces the security
guarantees (see documentation added in
188995648ac3adbb31212598afdbc4ae3c2c172f).

At this point, the CLI and aggregator still perform the strict
consistency check. The new default consistency check will be enabled in
v1.37 through a follow-up commit.