fix(node): ensure aggregate blob info is not deleted prematurely
There was a bug where in a particular corner case, an aggregate blob
info was deleted while some per-object blob info entries still existed
(see details below). If then the garbage collection attempted to update
the aggregate blob info for the expired object, it caused a panic in the
merge operator.

This fixes the bug by adding a check to delete the aggregate blob info
only if no corresponding per-object blob info entries still exist.

The bug was triggered in the `simtest_test_lagging_node_recovery` test
as follows:
- A blob was registered twice as deletable in epoch 2 with an end epoch
  of 4.
- One of the corresponding objects was deleted in epoch 3.
- The deletion event was processed at the beginning of epoch 4. At this
  point, the blob was already expired, meaning that blob data could be
  deleted.
- However, the second deletable object was still stored in the
  per-object blob info table as the clean-up was not yet complete.

Besides fixing the bug, this commit also adds a consistency check for
invariants of the blob info tables (turned off by default in the config)
and improves some logging.