refactor(core): prepare types for increase of max blob size
This commit prepares function and variable types for a future increase
of the blob size limit. Specifically, use `SymbolSizeType`, an alias of
`NonZeroU32`, instead of `NonZeroU16` to represent symbol sizes.

No behavior changes are introduced by this commit. An actual increase of
the maximum blob size requires related changes to the contracts, only
then can the limit be increased.

Also contains other minor cleanup and refactoring in the `walrus-core`
crate.