refactor: reduce database impact on recovery symbol calls
This commit improves the serving of recovery symbols in several ways.

First, it makes the fetching of the associated sliver an asynchronous operation so that it does not
block the core tokio thread. Second, fetching metadata is currently unnecessary since there is only
one supported encoding type, which can be known in advance. Third, requests to the blob-info table
which were performed once per symbol to check blob access is now performed once per request
(get-symbol or list-symbol).

Finally, whereas before it was possible that a single list-symbols request could have some symbols
fail to be fetched due to the symbol service being unavailable. Now, if the symbol service is ready
at the beginning of the request an attempt will be made to process all of the symbols.