refactor(sdk): use type-state pattern for blob store state machine (#2686)
This commit changes how we encode the state of to-be-stored blobs in the
client. Instead of using a single enum with a variant for each state, we
now use a type-state pattern with a separate type for each state.

This ensures that we always operate on blobs in the correct state, thus
simplifying some code and avoiding runtime panics. We currently keep an
owned reference to the blob object in the `UnencodedBlob` type, but this
will probably be changed to a `Bytes` object in a follow-up commit.

The majority of changes are in `crates/walrus-sdk/src/client.rs`,
`crates/walrus-sdk/src/client/client_types.rs`, and
`crates/walrus-sdk/src/client/resource.rs`, but there are related
changes in other locations as well.

All changes do not change any functionality, and are
backwards-compatible.