fix(sui): avoid redundant RPC requests for staking pools
The `get_committees_and_state` calls the `shard_assignment_to_committee`
multiple times for the different committees (present, past, and,
possibly, next).

Previously, fetching the `StakingPool` objects was done in this inner
function. As the different committees are almost identical in most
cases, this resulted in redundant RPC requests for the `StakingPool`
objects.

With this change, we fetch the `StakingPool` objects of all nodes that
are part of any committee once and pass them to the
`shard_assignment_to_committee` function. This ensures that each object
is only fetched once.

In addition, the requests are now performed concurrently using
`try_join_all`.

When running against Testnet, this reduces the duration of this
operation from ~800ms (or 1200ms if the next committee is also fetched)
to ~170ms.