feat(node): move garbage collection to a background task (#2725)
Moves the garbage collection added in previous commits to a background task. 

This background task is (by default) started at a pseudorandom time during the first half of an epoch, which is deterministically set based on the node's public key and the epoch. This is to make sure that not all nodes perform garbage collection at the same time.

Garbage collection is done in batches within blocking threads to make sure the normal node operation is not affected and to make sure that we can still interrupt the whole process and shut the node down cleanly even if GC is running.

The task is restarted upon node restart if it is not finished.