fix(node): avoid errors when storing metadata for untracked blobs
There is an extremely unlikely race condition when storing the blob
metadata where a blob-info entry is deleted between checking it in the
`StorageNodeInner::store_metadata` method and actually persisting the
metadata in the `Storage::put_metadata` method. This can result in a
`BlobInfoMergeOperand::MarkMetadataStored` merge operand being applied
to an untracked blob ID, which would previously result in a panic.

With this change, we allow the merge operation to proceed. The resulting
blob info will then be removed during the next garbage collection.

This commit also adds a new simtest where we repeatedly store and delete
blobs with a relatively short epoch duration.