chore: cherry-pick blob-info and garbage-collection improvements into the 1.38 release (#2757)
* Walrus v1.38.2 version bump

* fix(node): avoid errors when storing metadata for untracked blobs (#2747)

There is an extremely unlikely race condition when storing the blob
metadata where a blob-info entry is deleted between checking it in the
`StorageNodeInner::store_metadata` method and actually persisting the
metadata in the `Storage::put_metadata` method. This can result in a
`BlobInfoMergeOperand::MarkMetadataStored` merge operand being applied
to an untracked blob ID, which would previously result in a panic.

With this change, we allow the merge operation to proceed. The resulting
blob info will then be removed during the next garbage collection.

This commit also adds a new simtest where we repeatedly store and delete
blobs with a relatively short epoch duration.

* feat(typed-store): add snapshot functionality (#2739)

Also use it in blob-info consistency check.

* fix(node): delete per-object blob info when blob is deleted (#2746)

Also contains some test improvements:
* Improve the blob-info internal consistency check and fail in tests if
  the consistency check fails.
* Improve logging and error outputs in tests.